analyze/requirements.txt
asgiref==3.5.2
colorama==0.4.5
Django==3.2.12
django-cors-headers==3.13.0
django-extensions==3.2.1
djangorestframework==3.13.1
numpy==1.21.6
pandas==1.3.5
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
tqdm==4.64.1
typing_extensions==4.3.0
mysqlclient==1.4.2.post1
gunicorn==20.1.0
